fastapi_backend_service/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.32.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
alembic==1.14.0
aiosqlite==0.20.0
asyncpg==0.30.0          # PostgreSQL (production)

# ── Data Validation & Settings ────────────────────────────────────────────────
pydantic[email]==2.10.0
pydantic-settings==2.6.0

# ── Authentication ────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2            # passlib 1.7.4 incompatible with bcrypt >= 4.0.0

# ── AI / LLM ──────────────────────────────────────────────────────────────────
anthropic>=0.49.0        # Claude (Anthropic)
openai>=1.58.0           # Ollama / OpenAI-compatible backends (vLLM, LMStudio)
mem0ai>=0.1.29           # Long-term semantic memory

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx==0.28.0            # Async HTTP — MCP calls, event polling

# ── Scheduling ────────────────────────────────────────────────────────────────
apscheduler==3.11.0      # Auto-Patrol cron & event-driven scheduler

# ── Data Science (MCP sandbox) ────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.2.0
plotly>=5.24.0
scipy>=1.14.0
scikit-learn>=1.5.0

# ── Message Bus (optional — OOC event subscriber) ─────────────────────────────
nats-py>=2.9.0

# ── MongoDB (optional — OntologySimulator) ────────────────────────────────────
motor>=3.7.0
pymongo>=4.10.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
python-multipart==0.0.12

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0
